# Conflicts:
#	MSVC19/SocketBridge/SocketBridge.vcxproj.filters
#	MenuProgTranslator/translations.xlsx
#	QT-Linaro-Rockchip/SocketBridge/SocketBridge.pro
#	bin/current/da3/vmcDataFile.da3
#	bin/varie/prog/index_CN.html
#	bin/varie/prog/index_DE.html
#	bin/varie/prog/index_ES.html
#	bin/varie/prog/index_FR.html
#	bin/varie/prog/index_GB.html
#	bin/varie/prog/index_IT.html
#	bin/varie/prog/index_JP.html
#	bin/varie/prog/index_NL.html
#	bin/varie/prog/index_PL.html
#	bin/varie/prog/index_PT.html
#	bin/varie/prog/index_RO.html
#	bin/varie/prog/index_RU.html
#	bin/varie/prog/index_template.html
#	bin/varie/prog/js/Task_JP.js
#	src/CPUBridge/CPUBridge.cpp
#	src/CPUBridge/CPUBridge.h
#	src/CPUBridge/CPUBridgeEnumAndDefine.h
#	src/CPUBridge/CPUBridgeServer.cpp
#	src/CPUBridge/CPUBridgeServer.h
#	src/GPU/header.h
#	src/GPU/mainwindow.cpp
#	src/SocketBridge/CmdHandler/CmdHandler_ajaxReq_P0x25_caffeCortesia.cpp
#	src/SocketBridge/CmdHandler/CmdHandler_ajaxReq_P0x2B_notifyEndOfGrinderCleaningProc.cpp
#	src/SocketBridge/CmdHandler_ajaxReq.cpp
#	src/rheaCommonLib/OS/linux/linuxOS.cpp
#	src/rheaCommonLib/OS/linux/linuxOS.h
#	src/rheaCommonLib/OS/win/winOS.cpp
#	src/rheaCommonLib/OS/win/winOS.h
#	src/rheaCommonLib/rhea.h
#	src/rheaExternalSerialAPI/ESAPIProtocol.cpp

+++ bin/aaa_MH6_CREATE_rheAPI.sh
@@ -1,7 +1,7 @@
 clear
 DATA=`date '+%y%m%d'`
 estensione="_commit.mh6"
-filename="GPU_TS_v.2.5.1_$DATA$estensione"
+filename="GPU_TS_v.2.5.1-Beta_$DATA$estensione"
 
 
 FILE_GPU="./EMBEDDED_RELEASE_GPU"
@@ -18,27 +18,46 @@ if [ ! -f "$FILE_startRhea" ]; then
 	exit 1
 fi
 
+FILE_gpuUpdater="./EMBEDDED_RELEASE_GPUUpdater"
+if [ ! -f "$FILE_gpuUpdater" ]; then
+    echo "$FILE_gpuUpdater does not exist"
+	cp ./aaa_MH6_CREATE.sh ./error_EMBEDDED_RELEASE_GPUUpdate_does_not_exists
+	exit 1
+fi
+
+FILE_RSProto="./EMBEDDED_RELEASE_SecoBridge"
+if [ ! -f "$FILE_RSProto" ]; then
+    echo "$FILE_RSProto does not exist"
+	cp ./aaa_MH6_CREATE.sh ./error_EMBEDDED_RELEASE_SecoBridge_does_not_exists
+	exit 1
+fi
+
+
+
 #la roba che mi interessa la metto tutta in GPUPackage2019
 rm -r ./GPUPackage2019
 mkdir ./GPUPackage2019
 mkdir ./GPUPackage2019/current
 mkdir ./GPUPackage2019/current/lang
 cp "$FILE_GPU" ./GPUPackage2019/GPUFusion
+cp "$FILE_RSProto" ./GPUPackage2019/RSProto
 cp ./current/lang/*.* ./GPUPackage2019/current/lang
 cp ../src/makeRheaServicePack.sh ./GPUPackage2019
 rm ./varie/prog/lastUsedLang.txt
 cp -r ./varie ./GPUPackage2019
 chmod 777 ./GPUPackage2019/GPUFusion
+chmod 777 ./GPUPackage2019/RSProto
 chmod 777 ./GPUPackage2019/makeRheaServicePack.sh
 chmod 777 ./GPUPackage2019/current/lang
 rm ./GPUPackage2019/varie/prog/js/dev/compiler.jar
 
-#fuori da package c'e' solo startRhea.sh e il miniboot
+#fuori da package c'e' solo startRhea.sh e gpuUpdater
 cp "$FILE_startRhea" ./
 chmod 777 ./startRhea.sh
 
+
 rm ./$filename
-tar -czvf $filename ./GPUPackage2019 ./startRhea.sh
+tar -czvf $filename ./GPUPackage2019 ./startRhea.sh ./"$FILE_gpuUpdater"
 
 rm ./startRhea.sh
 rm -r ./GPUPackage2019
